+++ main/pool_state.h
@@ -161,10 +161,10 @@ typedef struct {
     uint8_t gateway_version_minor;
     bool gateway_version_valid;
 
-    // Temp Sensor firmware version
-    uint8_t temp_sensor_version_major;
-    uint8_t temp_sensor_version_minor;
-    bool temp_sensor_version_valid;
+    // Connect 8/10 Controller firmware version
+    uint8_t controller_version_major;
+    uint8_t controller_version_minor;
+    bool controller_version_valid;
 
     // Timers (up to MAX_TIMERS)
     timer_state_t timers[MAX_TIMERS];